pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["sql_lab*", "server*"]

[project]
name = "sql-lab"
version = "1.0.0"
description = "OpenEnv SQL Query Debugging & Optimization Lab"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "openai>=1.0.0",
]

[tool.uv]
dev-dependencies = []

[project.scripts]
server = "server.app:main"